+++ pkg/querier/distributor_queryable_test.go
@@ -369,3 +369,8 @@ func (m *mockDistributor) LabelNamesAndValues(ctx context.Context, matchers []*l
 	args := m.Called(ctx, matchers)
 	return args.Get(0).(*client.LabelNamesAndValuesResponse), args.Error(1)
 }
+
+func (m *mockDistributor) LabelValuesCardinality(ctx context.Context, labelNames []model.LabelName, matchers []*labels.Matcher) (uint64, *client.LabelValuesCardinalityResponse, error) {
+	args := m.Called(ctx, labelNames, matchers)
+	return args.Get(0).(uint64), args.Get(1).(*client.LabelValuesCardinalityResponse), args.Error(2)
+}